Only supply `NoOpAck` with explicit `@NonNull` param annotation.
