+++ spring-kafka/src/main/java/org/springframework/kafka/listener/adapter/KafkaBackoffAwareMessageListenerAdapter.java
@@ -50,8 +50,6 @@ public class KafkaBackoffAwareMessageListenerAdapter<K, V>
 		extends AbstractDelegatingMessageListenerAdapter<MessageListener<K, V>>
 		implements AcknowledgingConsumerAwareMessageListener<K, V> {
 
-	private static final Acknowledgment NO_OP_ACK = new NoOpAck();
-
 	private final String listenerId;
 
 	private final String backoffTimestampHeader;
@@ -96,12 +94,7 @@ public class KafkaBackoffAwareMessageListenerAdapter<K, V>
 				.ifPresent(nextExecutionTimestamp -> this.kafkaConsumerBackoffManager
 						.backOffIfNecessary(createContext(consumerRecord, nextExecutionTimestamp, consumer)));
 		try {
-			Acknowledgment ack = acknowledgment;
-			if (ack == null) {
-				// The default DLT handler now requires an Acknowledgment.
-				ack = NO_OP_ACK;
-			}
-			invokeDelegateOnMessage(consumerRecord, ack, consumer);
+			invokeDelegateOnMessage(consumerRecord, acknowledgment, consumer);
 		}
 		catch (Exception ex) {
 			throw new TimestampedException(ex, Instant.now(this.clock));
@@ -150,13 +143,4 @@ public class KafkaBackoffAwareMessageListenerAdapter<K, V>
 	public void onMessage(ConsumerRecord<K, V> data, Consumer<?, ?> consumer) {
 		onMessage(data, null, consumer);
 	}
-
-	static class NoOpAck implements Acknowledgment {
-
-		@Override
-		public void acknowledge() {
-		}
-
-	}
-
 }
