+++ spring-kafka/src/main/java/org/springframework/kafka/retrytopic/RetryTopicConfigurer.java
@@ -38,6 +38,7 @@ import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.kafka.support.EndpointHandlerMethod;
 import org.springframework.kafka.support.KafkaUtils;
 import org.springframework.kafka.support.TopicForRetryable;
+import org.springframework.lang.NonNull;
 import org.springframework.lang.Nullable;
 
 
@@ -469,7 +470,7 @@ public class RetryTopicConfigurer implements BeanFactoryAware {
 
 		public static final String DEFAULT_DLT_METHOD_NAME = "logMessage";
 
-		public void logMessage(Object message, @Nullable Acknowledgment ack) {
+		public void logMessage(Object message, @NonNull Acknowledgment ack) {
 			if (message instanceof ConsumerRecord) {
 				LOGGER.info(() -> "Received message in dlt listener: "
 						+ KafkaUtils.format((ConsumerRecord<?, ?>) message));
@@ -477,9 +478,7 @@ public class RetryTopicConfigurer implements BeanFactoryAware {
 			else {
 				LOGGER.info(() -> "Received message in dlt listener.");
 			}
-			if (ack != null) {
-				ack.acknowledge();
-			}
+			ack.acknowledge();
 		}
 
 	}
